requirements.txt
# Web Framework
Flask==3.1.2
flasgger==0.9.7.1

# Twilio SDK
twilio==9.9.0

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.45
psycopg[binary]==3.3.2

# Authentication
PyJWT==2.10.1
bcrypt==5.0.0

# Environment
python-dotenv==1.0.1

# Production Server (optional)
gunicorn==23.0.0
